подключил генерацию новых холмов и сделал очистку памяти

+++ ProjectGroup_Mars_Resource_Battle/MainProject/__astcache/d@@myprojects@mdndm@mdndm@projectgroup_mars_resource_battle@mainproject/classes@cell.h
@@ -79,7 +79,7 @@ public:
 
 private:
 	string type_cell;
-	float height_hill;
+	int height_hill;
 
 	rare rare_artefact;
 	usually usually_artefact;
